Add files via upload
Project: Live Weather Report (Text to Speech)
Problem: Trying to get the live update of the weather at current date and time when you plan to visit a destination city.
Approach: The code helps in providing the weather updates of any city in the world. It does the process by asking the user the required city and then extracting the live weather data from a web address using an API. The data further converted to a "( .json)" format and then transforms the required weather report to audio format. Speech helps the user to get the specific data info required.   
It helps the user with weather data and further plan their trip accordingly.